requirements.txt
openai==1.97.0
python-dotenv==1.1.1
tsplib95
